discover: align skill with file-based result channel
The score channel moved from stdout-parsing to $EVO_RESULT_PATH in
16b5845 (0.3.0). plugins/evo/skills/discover/ still described the
stdout contract: SKILL.md step 10a's output bullet, step 10b's "wire
protocol" sentence, the SDK and inline-helper reference docstrings,
and constructing-benchmark.md all said "stdout = single JSON with
score, stderr = everything else".

Step 10c's dry-run command piped stdout through validate_stdout.py.
That validator rejected any benchmark printing progress to stdout,
even though core.load_result ignores stdout when result.json exists.
Following the recipe against an SDK-instrumented benchmark with
normal logging would have failed validation.

- scripts/validate_result.py replaces validate_stdout.py: takes a
  path, asserts JSON object with numeric score. Mirrors the strict
  read in core.load_result.
- Step 10c allocates a fresh mktemp subdir per run. The helper's
  O_EXCL claim no longer collides on iteration; mirrors how evo run
  allocates attempts/NNN/.
- Reference docstrings drop the stdout-contract wording and note the
  stdout fallback for hand-run dev invocations.
