+++ plugins/evo/skills/discover/references/sdk_node.js
@@ -1,7 +1,8 @@
 // Node SDK usage example. Install: `npm install @evo-hq/evo-agent`.
 //
-// The SDK auto-reads $EVO_TRACES_DIR and $EVO_EXPERIMENT_ID. Traces flush
-// on each report() so the dashboard can stream progress live.
+// The SDK auto-reads $EVO_TRACES_DIR, $EVO_EXPERIMENT_ID, and
+// $EVO_RESULT_PATH. Traces flush on each report() so the dashboard can
+// stream progress live.
 
 import { Run, Gate } from '@evo-hq/evo-agent';
 
@@ -14,7 +15,8 @@ for (const task of tasks) {
   run.report(task.id, { score: result.score });
 }
 await run.finish();
-// finish(): prints score JSON to stdout, writes task_<id>.json per task.
+// finish(): writes score JSON to $EVO_RESULT_PATH (or stdout if unset)
+// and one task_<id>.json per task under $EVO_TRACES_DIR.
 
 // ---- Gate (exits 0 all-pass / 1 any-fail) ----
